Add StatsD metrics testing
Previously the `kinesis-it` project served as a single-purpose project
for handling kinesis-related it-testing. This change adds a new set of
integration specs for Statsd metrics. The project is therefore renamed.
In future we may choose to migrate to one `it` project per module, but
for now it does not really make sense.

Statsd metrics reporting is now verified through a dedicated IT test spinning
up a testcontainer.