Fix nacking unhandled pubsub messages during shutdown (#63)
The LowLevelSource maintains a list of what messages are in-flight but
not yet acked. During graceful shutdown, it should nack everything in
this list so that Pubsub can quickly re-send the messages to another
pod. But because of a subtle bug, the pubsub Subscriber was getting
shutdown before those messages got nacked.