Fix randomly adjusted first window size (#79)
common-streams has a feature where the first window is randomly adjusted
to a different size. This helps in a deployment where a large number of
pods might all start at the same time, but we pods to have mutually
staggered windows, e.g. to avoid write conflicts in the lake loader.

Unfortunately I broke the feature in #64 so this fixes it again.

I also add a test so it won't get broken again.