Refactor application health monitoring
Application health is needed for two distinct reasons:

1. Sending alerts to a monitoring webhook for setup errors. Relatedly,
   sending heartbeat messages to the webhook when healthy.
2. The HTTP health probe.  Needed so that the orchestration environment
   (kubernetes or whatever) can kill the pod when unhealthy

Several Snowplow apps each implement their own logic of when to toggle
the health probe and when to send an alert.  This PR consolidates that
logic into one place.

After this PR, the application code just needs to call methods on the
AppHealth class.  This lib then manages webhook events and health probe,
based on the current status of the AppHealth.