Resolve schemas in parallel (#85)
I have seen examples from our apps where cpu usage and event throughput
periodically drops.  This appears to coincide with the Iglu cache
expiration time.

I believe this happens because all schemas tend to expire at the same
time and need to be re-fetched by iglu-scala-client. Currently, we
traverse over schemas sequentially, so we need to wait for each success
before fetching the next schema.  For a pipeline using many schemas,
this can be a long period of downtime (several seconds) as we pause for
schema resolution.

This commit changes to resolving schemas in parallel, so the downtime
pauses should be shorter.