Pubsub source open more transport channels (#92)
Previously the pubsub SDK opened just a single transport channel
(roughly equivalent to a TCP channel).

Some snowplow apps we need to batch up a very large number of events
from the source before acking.  In those circumstances it is good to
open more streaming pulls (a type of RPC) to keep the pubsub
subscription healthy.  But to open more streaming pulls we also need to
open more transport channels (TCP channels) to support all the RPCs.

This commit adds a new config parameter `maxPullsPerTransportChannel`.
The default value 16 is good for most apps.