Fix for `_schema_version` overriding non-object entity data (#97)
We recommend that Snowplow entities should be JSON *objects* only. But
technically it is possible for them to be something else, e.g. a JSON
array.

Before this fix, we accidentally erased the original value if a context
was not a JSON object. The bug only affected unstructed destinations,
e.g. the Snowflake loader.

The fix was implented in snowplow/snowplow-scala-analytics-sdk#137 and
is imported here.