Sources should report stream latency of stuck events (#105)
Currently in common-streams apps, the _application_ (i.e. not this
library) is responsible for tracking latency of events consumed from the
stream.  Because of the way it is implemented in the apps, if an event
gets stuck (e.g. cannot be written to destination) then the latency
drops to zero for the period when the app is retrying the stuck event.

This commit aims to fix the problem where the `latency_millis` metric
wrongly drops to zero:

1. The `SourceAndAck` is now responsible for tracking the latency metric
and pushing the metric to the app. This consolidates the metric across
multiple apps into one place

2. The Metrics class allows metric starting values to be wrapped in and
Effect, i.e. `F[Metrics.State]`. This means the application can pass in
the latency reported by the Source.

3. The `TokenedEvents` now does not contain a stream timestamp.  The
downstream app has no business knowing the stream timestamp, now that
latency is calculated inside common-streams lib.

4. The `LowLevelSource` no longer provides a `lastLiveness`.  Instead,
the low level source is expected to periodically emit a `None` whenever
it is healthy. The higher level `SourceAndAck` then takes responsibility
for monitoring whether the low level source is healthy.

5. As a result of these changes, the `SourceAndAck` emits a latency
metric of zero whenever the source is genuinely healthy but there are no
events to be pulled. This means the app's latency metric is only zero if
there are no events.  This is better than the previous situation, where
the app might mis-leadingly emit a zero latency when the app is stuck,
e.g. cannot connect to warehouse.