Sources should report stream latency of stuck events
Currently in common-streams apps, the _application_ (i.e. not this
library) is responsible for tracking latency of events consumed from the
stream.  Because of the way it is implemented in the apps, if an event
gets stuck (e.g. cannot be written to destination) then the latency
drops to zero for the period when the app is retrying the stuck event.

This commit aims to fix the problem where the `latency_millis` metric
wrongly drops to zero.

1. The source itself now tracks the latency of the event which is
   currently being processed by the downstream application.
2. The Metrics class allows metric starting values to be wrapped in and
   Effect, i.e. `F[Metrics.State]`. This means the application can pass
   in the latency reported by the Source.