Sources should report stream latency of stuck events - amendments 1
The `SourceAndAck` is now responsible for pushing the latency metric.  This
consolidates the metric across multiple apps into one place.

The `TokenedEvents` now does not contain a stream timestamp.  The downstream
app has no business knowing the stream timestamp, now that latency is
calculated inside common-streams lib.

The `LowLevelSource` no longer provides a `lastLiveness`.  Instead, the low
level source is expected to periodically emit a `None` whenever it is
healthy. The higher level `SourceAndAck` then takes responsibility for
monitoring whether the low level source is healthy.

As a result of these changes, the `SourceAndAck` emits a latency metric
of zero whenever the source is genuinely healthy but there are no events
to be pulled. This means the app's latency metric is only zero if there
are no events.  This is better than the previous situation, where the
app might mis-leadingly emit a zero latency when the app is stuck, e.g.
cannot connect to warehouse.