Ensure first window is always small (#106)
This change affects windowing apps e.g. Lake Loader.

Before this change, the first window was a random size, to avoid write
conflicts between several loaders finishing their windows at similar
times. After this change, the first window is still random, but sized in
the ranged 25% to 50% of the regular window size. This was shown in Lake
Loader to give a smoother starting cpu profile, and it prevents the
loader from pulling in too many events into the first window.