Debounce how often we checkpoint progress (#109)
In snowplow-incubator/snowflake-loader#57 we added code to the snowflake
loader so it checkpoints once every 10 seconds instead of once per
batch. This meant we could decrease the write-throughput requirements of
the DynamoDB table.

This commit moves the logic over here into common-streams so that all
loaders get the benefit of this improvement.