Prefetch from pubsub source when parallelPullCount is 1 (#110)
The pubsub source uses `parEvalMapUnordered` when pulling from pubsub.  If parallelism is >1 then this has the effect of pre-fetching events, so the downstream app never needs to wait for new events.  But if parallelism is 1, then `parEvalMapUnordered` does not do any pre-fetching.

This PR changes things so if parallelism is 1, then we explicitly prefetch a batch.  This means the pubsub source behaves more consistently across different parallelisms, and more similar to the other sources.