Pubsub sink elimate Publisher wrapper (#121)
The Pubsub `Sink` used the 3rd party `Publisher` class to publish events
to the topic. But this was messy in common-streams, because the
`Publisher` is mainly there to batch up events into bigger requests,
whereas common-streams apps already operate in batches so we don't need
that help.

This commit eliminates the `Publisher` and insteads uses the underlying
GRPC stub directly.