Add new method `pingForHealth` to sinks
We want to start using the common-streams sinks in the snowplow
collector. Until now, common-streams only discovers a sink is
unavailable if it fails to send events to the sink. But the collector
has a requirement that it needs to know the health of the sink
immediately upon startup, see snowplow/stream-collector#276

This commit adds a new method `pingForHealth` to the `Sink` trait.
Implementations of `Sink` should probe the external resource to check
its health.