Fix for blocked thread in Kafka sink
In #120 we introduced a single-threaded executor for blocking operations
in the kafka producer. But I was mistakenly using the same thread for
awaiting for confirmation the events got sent to the kafka server. This
caused a blocked thread, which prevented more events from getting sent.

This fix works by adding a second thread for awaiting for events to get
sent to the kafka server.