Kinesis Source: Change from SynchronousQueue to CountdownLatch for back pressure
This fixes a problem where an app could OOM when Kinesis scales to add
more shards. It relates to the feature implemented in
snowplow-incubator/common-streams#102.

It is a requirement of the Source that the ShardRecordProcessor gets
blocked until the downstream app is ready to consume an event. Before
this PR we used a SynchronousQueue to achieve this blocking. After this
PR we instead use a CountdownLatch, plus an unbounded queue.

This means we have better control over backpressure during the scenario
where the Source tries to handle many shard ends at the same time.